chore(react): move webchatReducer to its own file
