+++ packages/botonic-react/src/webchat/hooks.js
@@ -1,6 +1,7 @@
-import { useEffect, useReducer, useState, useRef } from 'react'
-import { WEBCHAT, COLORS } from '../constants'
+import { useEffect, useReducer, useRef, useState } from 'react'
+import { COLORS, WEBCHAT } from '../constants'
 import { scrollToBottom } from '../utils'
+import { webchatReducer } from './webchat-reducer'
 
 export const webchatInitialState = {
   width: WEBCHAT.DEFAULTS.WIDTH,
@@ -29,72 +30,6 @@ export const webchatInitialState = {
   error: {},
   devSettings: {},
   isWebchatOpen: false,
-}
-
-export function webchatReducer(state, action) {
-  switch (action.type) {
-    case 'addMessage':
-      if (
-        state.messagesJSON &&
-        state.messagesJSON.find(m => m.id === action.payload.id)
-      )
-        return state
-      return {
-        ...state,
-        messagesJSON: [...(state.messagesJSON || []), { ...action.payload }],
-      }
-    case 'addMessageComponent':
-      return {
-        ...state,
-        messagesComponents: [...state.messagesComponents, action.payload],
-      }
-    case 'updateMessage':
-      const msgIndex = state.messagesJSON
-        .map(m => m.id)
-        .indexOf(action.payload.id)
-      if (msgIndex > -1)
-        return {
-          ...state,
-          messagesJSON: [
-            ...state.messagesJSON.slice(0, msgIndex),
-            { ...action.payload },
-            ...state.messagesJSON.slice(msgIndex + 1),
-          ],
-        }
-      return state
-    case 'updateReplies':
-      return { ...state, replies: action.payload }
-    case 'updateLatestInput':
-      return { ...state, latestInput: action.payload }
-    case 'updateTyping':
-      return { ...state, typing: action.payload }
-    case 'updateWebview':
-      return { ...state, ...action.payload }
-    case 'updateSession':
-      return { ...state, session: { ...action.payload } }
-    case 'updateUser':
-      return { ...state, user: { ...action.payload } }
-    case 'updateLastRoutePath':
-      return { ...state, lastRoutePath: action.payload }
-    case 'updateHandoff':
-      return { ...state, handoff: action.payload }
-    case 'updateTheme':
-      return { ...state, theme: { ...action.payload } }
-    case 'updateDevSettings':
-      return { ...state, devSettings: { ...action.payload } }
-    case 'toggleWebchat':
-      return { ...state, isWebchatOpen: action.payload }
-    case 'setError':
-      return { ...state, error: action.payload || {} }
-    case 'clearMessages':
-      return {
-        ...state,
-        messagesJSON: [],
-        messagesComponents: [],
-      }
-    default:
-      throw new Error()
-  }
 }
 
 export function useWebchat() {
